Added checking for firmware
