+++ Hspi/OpenZWaveDB/ZWaveDeviceManufacturer.cs
@@ -1,0 +1,9 @@
+﻿#nullable enable
+
+namespace Hspi.OpenZWaveDB
+{
+    internal record ZWaveDeviceManufacturer
+    {
+        public string? Label { get; init; }
+    }
+}
