+++ Hspi/OpenZWaveDB/ZWaveDeviceParameterOption.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+using static System.FormattableString;
+
+#nullable enable
+
+namespace Hspi.OpenZWaveDB
+{
+    internal record ZWaveDeviceParameterOption
+    {
+        public string? Label { get; init; }
+        public int Value { get; init; }
+
+        [JsonIgnore]
+        public string Description => Invariant($"{Value} - {Label}");
+    }
+}
